MB-54801: Avoid purging stale items while referenced by older items
In ephemeral buckets, old versions of documents may briefly be retained
in memory (in the sequence list) if they are required by an ongoing
backfill.

Stale items may have a "replacement" pointer set, pointing
to the newer version of the value.

This is used to determine if the replacement appears in the range of
seqnos the backfill covers. This prevents multiple versions of a value
being sent in a given backfill.  Stale items can't be unconditionally
skipped, as that could lead to _no_ version of a value being sent in a
backfill.

The pointed-to value can itself be made stale if the document is
updated again. I.e.,:

1.  V1(New)

2.  V1(Stale) -> V2(New)

3.  V1(Stale) -> V2(Stale) -> V3(New)

This is fine. _However_, the EphTombstoneStaleItemDeleter may pause,
releasing its range lock (to avoid blocking backfills for a long time).
If it happened to pause between the seqnos of V1 and V2, once it
resumes it may purge V2:

1. Item stored
    V1(New) <other items>

2. Purger runs, pauses _after_ V1
    V1(New) <other items>
             ^
             paused here

3. Item updated
    V1(Stale) <other items> V2(New)
              ^

4. Item updated
    V1(Stale) <other items> V2(Stale) V3(New)
              ^

5. Purger resumes, purges V2
    V1(Stale) <other items> XXXXXX V3(New)

This would leave V1 with a dangling replacement ptr:

V1(Stale) -> XXXXX      ... V3(New)

This is unsafe; a subsequent backfill would read from garbage memory,
potentially leading to a crash or an incorrect backfill.

In _principle_, this could be fixed by updating the replacement ptr to
point V1->V3.
This might appear to risk consistency issues e.g., "what if a new
backfill covered seqnos including V1 and V2, but not V3? It would
incorrectly see an older version V1, the document would appear to have
reverted". However, that particular scenario would not be possible;
backfills will always end at the high seqno at the time of their
creation; a new backfill would always include V3 as well.

Note that it would not be correct to set the replacement ptr of V1 back
to a nullptr when V2 is purged; that could lead to V1 and V3 both
appearing in a backfill.

Stale items can not be purged while there are ongoing backfills
locking a range of seqnos including the stale item, so there is no
concern for existing backfills.

However, locating the _previous_ version of a document in the seqlist
would be costly; the seqlist is a singly linked list, and we do not
record a pointer "back" to the previous version, only from the stale
item to it's replacement.

A simpler approach (which is the one taken in this patch) is to forbid
purging stale items while an older stale version still points to them.

This patch achieves this by adding a bit flag indicating if _this_
OrderedStoredValue is the replacement for an older version.  This is
set when the new OSV is appended to the seqlist, and is cleared when
the older stale version is purged (the old version holds a pointer to
its replacement, so it is easy to find the new value and clear this
flag).

In the above case, that would make V2 ineligible for purging until
after V1 is purged; this ensures versions of a value are purged in
the order they were created.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/185254
Well-Formed: Restriction Checker
Tested-by: James H <james.harrison@couchbase.com>

+++ engines/ep/src/linked_list.cc
@@ -263,10 +263,12 @@ size_t BasicLinkedList::purgeTombstones(
         }
 
         bool replaced = false;
+        bool isReplacement = false;
         {
             std::lock_guard<std::mutex> writeGuard(getListWriteLock());
             stale = it->isStale(writeGuard);
             replaced = it->getReplacementIfStale(writeGuard);
+            isReplacement = it->isStaleReplacement();
         }
 
         bool isDropped = false;
@@ -274,8 +276,10 @@ size_t BasicLinkedList::purgeTombstones(
             isDropped = isDroppedKeyCb(it->getKey(), it->getBySeqno());
         }
 
-        // Only stale or dropped items are purged.
-        if (stale || isDropped) {
+        // Only stale or dropped items are purged, but they _cannot_ be
+        // purged if they are pointed-to by an even older stale version of
+        // the same value which has not yet been purged
+        if (!isReplacement && (stale || isDropped)) {
             // Checks pass, remove from list and delete.
             it = purgeListElem(it, stale, replaced);
             ++purgedCount;
